+++ kadai-adapter-camunda-8-system-connector/src/main/java/io/kadai/adapter/systemconnector/camunda/tasklistener/UserTaskCompletion.java
@@ -2,14 +2,11 @@ package io.kadai.adapter.systemconnector.camunda.tasklistener;
 
 import io.camunda.client.api.response.ActivatedJob;
 import io.camunda.spring.client.annotation.JobWorker;
-import io.kadai.adapter.exceptions.TaskTerminationFailedException;
-import io.kadai.adapter.impl.scheduled.UserContext;
 import io.kadai.adapter.impl.service.KadaiTaskCompletionService;
 import io.kadai.adapter.systemconnector.api.ReferencedTask;
 import io.kadai.adapter.systemconnector.camunda.tasklistener.util.ReferencedTaskCreator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -20,9 +17,6 @@ public class UserTaskCompletion {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(UserTaskCompletion.class);
   private boolean gotActivated = false;
-
-  @Value("${kadai.adapter.run-as.user}")
-  protected String runAsUser;
 
   public UserTaskCompletion(
       KadaiTaskCompletionService taskTerminator, ReferencedTaskCreator referencedTaskCreator) {
@@ -46,16 +40,7 @@ public class UserTaskCompletion {
 
       ReferencedTask referencedTask = referencedTaskCreator.createReferencedTaskFromJob(job);
       // todo: system url is missing here
-      UserContext.runAsUser(
-          runAsUser,
-          () -> {
-            try {
-              taskTerminator.terminateKadaiTask(referencedTask);
-            } catch (TaskTerminationFailedException e) {
-              throw new RuntimeException(e);
-            }
-            return null;
-          });
+      taskTerminator.terminateKadaiTask(referencedTask);
 
     } catch (Exception e) {
       LOGGER.error(
