+++ kadai-adapter-camunda-8-system-connector/src/main/java/io/kadai/adapter/systemconnector/camunda/tasklistener/UserTaskCreation.java
@@ -2,14 +2,11 @@ package io.kadai.adapter.systemconnector.camunda.tasklistener;
 
 import io.camunda.client.api.response.ActivatedJob;
 import io.camunda.spring.client.annotation.JobWorker;
-import io.kadai.adapter.exceptions.TaskCreationFailedException;
-import io.kadai.adapter.impl.scheduled.UserContext;
 import io.kadai.adapter.impl.service.KadaiTaskStarterService;
 import io.kadai.adapter.systemconnector.api.ReferencedTask;
 import io.kadai.adapter.systemconnector.camunda.tasklistener.util.ReferencedTaskCreator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -20,9 +17,6 @@ public class UserTaskCreation {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(UserTaskCreation.class);
   private boolean gotActivated = false;
-
-  @Value("${kadai.adapter.run-as.user}")
-  protected String runAsUser;
 
   public UserTaskCreation(
       KadaiTaskStarterService taskStarter,
@@ -47,16 +41,7 @@ public class UserTaskCreation {
 
       ReferencedTask referencedTask = referencedTaskCreator.createReferencedTaskFromJob(job);
 
-      UserContext.runAsUser(
-          runAsUser,
-          () -> {
-            try {
-              taskStarter.createKadaiTask(referencedTask);
-            } catch (TaskCreationFailedException e) {
-              throw new RuntimeException(e);
-            }
-            return null;
-          });
+      taskStarter.createKadaiTask(referencedTask);
 
     } catch (Exception e) {
       LOGGER.error(
