+++ kadai-adapter/src/main/java/io/kadai/adapter/impl/service/KadaiTaskCompletionServiceImpl.java
@@ -19,10 +19,13 @@
 package io.kadai.adapter.impl.service;
 
 import io.kadai.adapter.exceptions.TaskTerminationFailedException;
+import io.kadai.adapter.impl.util.UserContext;
 import io.kadai.adapter.manager.AdapterManager;
 import io.kadai.adapter.systemconnector.api.ReferencedTask;
+import io.kadai.common.internal.util.CheckedSupplier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -33,6 +36,9 @@ public class KadaiTaskCompletionServiceImpl implements KadaiTaskCompletionServic
 
   private final AdapterManager adapterManager;
 
+  @Value("${kadai.adapter.run-as.user}")
+  protected String runAsUser;
+
   public KadaiTaskCompletionServiceImpl(AdapterManager adapterManager) {
     this.adapterManager = adapterManager;
   }
@@ -42,7 +48,15 @@ public class KadaiTaskCompletionServiceImpl implements KadaiTaskCompletionServic
       throws TaskTerminationFailedException {
     LOGGER.trace("KadaiTaskCompletionService.terminateKadaiTask ENTRY");
 
-    adapterManager.getKadaiConnector().terminateKadaiTask(referencedTask);
+    UserContext.runAsUser(
+        runAsUser,
+        CheckedSupplier.rethrowing(
+            () -> {
+              adapterManager.getKadaiConnector().terminateKadaiTask(referencedTask);
+              return null;
+            }
+        )
+    );
 
     LOGGER.trace("KadaiTaskCompletionService.terminateKadaiTask EXIT");
   }
