+++ kadai-adapter/src/main/java/io/kadai/adapter/impl/service/KadaiTaskStarterServiceImpl.java
@@ -19,12 +19,15 @@
 package io.kadai.adapter.impl.service;
 
 import io.kadai.adapter.exceptions.TaskCreationFailedException;
+import io.kadai.adapter.impl.util.UserContext;
 import io.kadai.adapter.kadaiconnector.api.KadaiConnector;
 import io.kadai.adapter.manager.AdapterManager;
 import io.kadai.adapter.systemconnector.api.ReferencedTask;
+import io.kadai.common.internal.util.CheckedSupplier;
 import io.kadai.task.api.models.Task;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -33,6 +36,9 @@ public class KadaiTaskStarterServiceImpl implements KadaiTaskStarterService {
   private static final Logger LOGGER = LoggerFactory.getLogger(KadaiTaskStarterServiceImpl.class);
 
   private final AdapterManager adapterManager;
+
+  @Value("${kadai.adapter.run-as.user}")
+  protected String runAsUser;
 
   public KadaiTaskStarterServiceImpl(AdapterManager adapterManager) {
     this.adapterManager = adapterManager;
@@ -45,8 +51,16 @@ public class KadaiTaskStarterServiceImpl implements KadaiTaskStarterService {
 
     KadaiConnector kadaiConnector = adapterManager.getKadaiConnector();
 
-    Task kadaiTask = kadaiConnector.convertToKadaiTask(referencedTask);
-    kadaiConnector.createKadaiTask(kadaiTask);
+    UserContext.runAsUser(
+        runAsUser,
+        CheckedSupplier.rethrowing(
+            () -> {
+              Task kadaiTask = kadaiConnector.convertToKadaiTask(referencedTask);
+              kadaiConnector.createKadaiTask(kadaiTask);
+              return null;
+            }
+        )
+    );
 
     LOGGER.trace("KadaiTaskStarterService.createKadaiTask EXIT");
   }
